refactor(db): add force option to the databases
rename some files
