+++ src/files/eshopdb/index.ts
@@ -2,16 +2,20 @@ import { getDataDir } from "../../config";
 import { create } from "../../logger";
 import { EShopDB } from "./eshopdb";
 
-const cachedDb = new EShopDB();
+let cachedDb: EShopDB;
 
-export async function getEShopDB(dataDir: string = getDataDir()) {
+export async function getEShopDB({
+  force = false,
+  dataDir = getDataDir()
+}: Partial<DBOptions> = {}) {
   const log = create("eshopdb:init");
-  if (!cachedDb.isOutdated()) {
+  if (cachedDb && !force && !cachedDb.isOutdated()) {
     log.debug("Using cached version of EShopDB");
     return cachedDb;
   }
 
   log.debug("Init EShopDB");
+  cachedDb = new EShopDB();
   cachedDb.setOutputDir(dataDir);
   return cachedDb.initDb();
 }
