+++ src/files/nswdb/index.ts
@@ -2,19 +2,20 @@ import { getDataDir, INSWDBOptions } from "../../config";
 import { create } from "../../logger";
 import { NSWDB } from "./nswdb";
 
-const cachedDb: NSWDB = new NSWDB();
+let cachedDb: NSWDB;
 
 export async function getNSWDB(
   dataDir: string = getDataDir(),
   { force = false, refreshInterval }: INSWDBOptions = {}
 ) {
   const log = create("nswdb:init");
-  if (!force && !cachedDb.isOutdated(refreshInterval)) {
+  if (cachedDb && !force && !cachedDb.isOutdated(refreshInterval)) {
     log.debug("Using cached version of NSWDB");
     return cachedDb;
   }
 
   log.debug("Init NSWDB");
+  cachedDb = new NSWDB();
   cachedDb.setOutputDir(dataDir);
   return cachedDb.initDb(force);
 }
