+++ src/files/thegamesdb/index.ts
@@ -4,9 +4,12 @@ import { TGDB } from "./tgdb";
 
 const cachedDb = new TGDB();
 
-export async function getTGDB(dataDir: string = getDataDir()) {
+export async function getTGDB({
+  force = false,
+  dataDir = getDataDir()
+}: Partial<DBOptions> = {}) {
   const log = create("tgdb:init");
-  if (!cachedDb.isOutdated()) {
+  if (!force && !cachedDb.isOutdated()) {
     log.debug("Using cached version of TGDB");
     return cachedDb;
   }
