+++ src/types/jsondb.d.ts
@@ -1,0 +1,4 @@
+interface DBOptions {
+  force: boolean;
+  dataDir: string;
+}
